feat(code): create initial framework for bytecodes (#1)
The initial implementation introduces the definitions for the machine
opcodes and some functions for converting instructions.  To start, only
OpConstant is defined.  Additional instructions will be added
iteratively.

The unit test validates that creation of an OpConstant with a two-byte
wide operand can be successfully created and encoded in big-endian
notation.