feat(stack-cleanup): do not leave the expression results on the stack
Expressions should not be reused by definition.  Thus, keeping the value
on the stack is wrong and will lead to eventual stack overflow.

It was originally implemented with the involuntary stack persistence for
early development and testing.  This is because we did not yet have an
instruction for popping a value from the stack.